added rudimentary handling for notify messages

+++ chaninterface.go
@@ -82,6 +82,14 @@ func (c *chaninterface) OnMessage(address, message string) {
 				return
 			}
 			c.handlePushMessage(address, msg)
+		case shared.MsgNotify:
+			msg := &shared.NotifyMessage{}
+			err := json.Unmarshal([]byte(message), msg)
+			if err != nil {
+				log.Println("OnMessage: failed to parse JSON!", err)
+				return
+			}
+			c.handleNotifyMessage(address, msg)
 		default:
 			log.Println("OnMessage: WARNING: Unknown object received:", msgType.String())
 		}
@@ -140,11 +148,15 @@ OnFileReceived is called when a file has been successfully received.
 */
 func (c *chaninterface) OnFileReceived(address, path, name string) {
 	// TODO fix this: NOTE: no lock check so that locks don't have to stay on for long file transfers
+	// FIXME debug: don't remove peer files until we know why wrong data is written to them. Check if temp data is already wrong or not.
+	removeTemp := true
 	// no matter what, remove temp file
 	defer func() {
-		err := os.Remove(path)
-		if err != nil {
-			log.Println("OnFileReceived: failed to remove temp file:", err)
+		if removeTemp {
+			err := os.Remove(path)
+			if err != nil {
+				log.Println("OnFileReceived: failed to remove temp file:", err)
+			}
 		}
 		// remove from allowedTransfers
 		c.mutex.Lock()
@@ -173,6 +185,7 @@ func (c *chaninterface) OnFileReceived(address, path, name string) {
 		err = ioutil.WriteFile(path, data, shared.FILEPERMISSIONMODE)
 		// log.Println("DEBUG: wrote model.")
 	case shared.OtPeer:
+		removeTemp = false
 		// peers are written to disk too, but in correct dir with pm.Name
 		path := c.enc.RootPath + "/" + shared.ORGDIR + "/" + shared.PEERSDIR + "/" + pm.Identification
 		err = ioutil.WriteFile(path, data, shared.FILEPERMISSIONMODE)
